+++ ms-historial/src/main/java/com/ticketuki/historialservice/repository/HistorialRepository.java
@@ -13,10 +13,10 @@ import java.util.List;
 public interface HistorialRepository extends JpaRepository<Historial, Long> {
 
     @Query("SELECT h FROM Historial h WHERE h.entidad = :entidad AND h.id_entidad = :idEntidad")
-    List<Historial> findByEntidadAndId_entidad(@Param("entidad") TipoEntidad entidad, @Param("idEntidad") Integer idEntidad);
+    List<Historial> findByEntidadAndId_entidad(@Param("entidad") TipoEntidad entidad, @Param("idEntidad") Long idEntidad);
 
     @Query("SELECT h FROM Historial h WHERE h.usuario_id = :usuarioId")
-    List<Historial> findByUsuario_id(@Param("usuarioId") Integer usuarioId);
+    List<Historial> findByUsuario_id(@Param("usuarioId") Long usuarioId);
 
     List<Historial> findByTimestampBetween(LocalDateTime inicio, LocalDateTime fin);
 }
